+++ spec/services/queries/daily_metrics_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Queries::DailyMetrics do
+RSpec.describe Metrics::Group::ByDay do
   describe '.call' do
     let(:project) { create(:project, name: 'rs-code-review-metrics') }
